Slight improvement(client)
1. package name com.testbuild to com.bookaholic
2. format codes , clean codes
3. clear minor errors
4. make compilable and runnable
5. authentication
6. in XmlHandlerRawItem,ArrayDeqeue to Stack for Android platform 8
compatibility

+++ client/MainActivityII/src/com/fima/cardsui/Utils.java
@@ -4,8 +4,6 @@ import android.content.Context;
 import android.util.DisplayMetrics;
 
 public class Utils {
-	
-	
 
 	/**
 	 * This method converts device specific pixels to device independent pixels.
@@ -29,11 +27,11 @@ public class Utils {
 		int px = (int) (dp * (metrics.densityDpi / 160f));
 		return px;
 	}
-	
+
 	public static float convertDpToPixel(Context context, float dp) {
 
 		DisplayMetrics metrics = context.getResources().getDisplayMetrics();
-		float px = (float) (dp * (metrics.densityDpi / 160f));
+		float px = (dp * (metrics.densityDpi / 160f));
 		return px;
 	}
